+++ Engine/Source/GUI/widget_render_object.h
@@ -14,8 +14,8 @@ namespace Ming3D
         ERenderType mRenderType = ERenderType::Transparent;
 
     public:
-        virtual int GetNumBatches() override;
-        virtual void GetRenderBatch(int batchIndex, RenderBatch* outBatch) override;
+        virtual size_t GetNumBatches() override;
+        virtual void GetRenderBatch(size_t batchIndex, RenderBatch* outBatch) override;
 
         virtual ERenderType GetRenderType() override;
         virtual bool GetCastShadows() override;
