+++ Engine/Source/SceneRenderer/render_object.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "SceneRenderer/render_type.h"
 #include "glm/vec3.hpp"
+#include <cstddef>
 
 namespace Ming3D
 {
@@ -12,9 +13,9 @@ namespace Ming3D
         virtual ~RenderObject() {}
 
         /* Gets the number of render batches in this object. */
-        virtual int GetNumBatches() = 0;
+        virtual size_t GetNumBatches() = 0;
         /* Gets the specified render batch (at the specified index). */
-        virtual void GetRenderBatch(int batchIndex, RenderBatch* outBatch) = 0;
+        virtual void GetRenderBatch(size_t batchIndex, RenderBatch* outBatch) = 0;
 
         virtual ERenderType GetRenderType() = 0;
